backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0

# OCR
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# ML/NLP
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2

# Database
psycopg2-binary==2.9.9
